requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-import-export==2.0.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
PyYAML==5.3.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
